lab3-FSI-LLM-Demo/requirements.txt
pyyaml
langchain
bs4
lxml
tqdm
regex
opensearch-py
boto3==1.26.73
pandas==1.5.3
Pillow==9.4.0
streamlit==1.20.0
# torch==1.13.1
# torchvision==0.14.1
requests-aws4auth